- Split the api out to public and private routes - Pushed the middleware back into server.js
